+++ twine/commands/register.py
@@ -62,18 +62,24 @@ def main(args):
     parser = argparse.ArgumentParser(prog="twine register")
     parser.add_argument(
         "-r", "--repository",
+        action=utils.EnvironmentDefault,
+        env='TWINE_REPOSITORY',
         default="pypi",
         help="The repository to register the package to (default: "
              "%(default)s)",
     )
     parser.add_argument(
-        "-u", "--username", action=utils.EnvDefault, env='PYPI_USER',
+        "-u", "--username",
+        action=utils.EnvironmentDefault,
+        env='TWINE_USERNAME',
         required=False, help="The username to authenticate to the repository "
                              "as (can also be set via %(env)s environment "
                              "variable)",
     )
     parser.add_argument(
-        "-p", "--password", action=utils.EnvDefault, env='PYPI_PASSWORD',
+        "-p", "--password",
+        action=utils.EnvironmentDefault,
+        env='TWINE_PASSWORD',
         required=False, help="The password to authenticate to the repository "
                              "with (can also be set via %(env)s environment "
                              "variable)",