+++ twine/commands/upload.py
@@ -149,6 +149,8 @@ def main(args):
     parser = argparse.ArgumentParser(prog="twine upload")
     parser.add_argument(
         "-r", "--repository",
+        action=utils.EnvironmentDefault,
+        env="TWINE_REPOSITORY",
         default="pypi",
         help="The repository to upload the files to (default: %(default)s)",
     )
@@ -168,13 +170,17 @@ def main(args):
         help="GPG identity used to sign files",
     )
     parser.add_argument(
-        "-u", "--username", action=utils.EnvDefault, env='PYPI_USER',
+        "-u", "--username",
+        action=utils.EnvironmentDefault,
+        env="TWINE_USERNAME",
         required=False, help="The username to authenticate to the repository "
                              "as (can also be set via %(env)s environment "
                              "variable)",
     )
     parser.add_argument(
-        "-p", "--password", action=utils.EnvDefault, env='PYPI_PASSWORD',
+        "-p", "--password",
+        action=utils.EnvironmentDefault,
+        env="TWINE_PASSWORD",
         required=False, help="The password to authenticate to the repository "
                              "with (can also be set via %(env)s environment "
                              "variable)",